Switch SSM runtime to float32 and add a streaming dSMC combine kernel
  - Drop forced jax_enable_x64; run SSM prior/runtime/parameterization in
    float32/int32 for GPU throughput
  - Add `_dsmc_combine_kernel` Pallas/Triton kernel that computes the seam
    logsumexp without materializing the (P,N,N) tensor (tf32 cross on GPU,
    exact-fp32 interpret on CPU/tests); rewrite dSMC `_stitch_logits` to use it
  - Extract SSM construction helpers from `tests/ssm_test_utils` into
    `nof1_causal_lab.models.ssm.testing` so benchmark scripts can reuse them
  - Replace `scripts/benchmark.py`/`profile_l4.py` with a `scripts/benchmarks/`
    suite (GPU + synthetic-nonlinear) plus `scripts/pallas_combine.py`
  - Refresh inference docs to `marginal_particle_gibbs`/`particle_marginal_mh`
    naming and corrected compilation paths

+++ apps/data-pipeline/notebooks/samplers.py
@@ -90,7 +90,7 @@ def run_pathfinder(target: Target, config: PathfinderConfig | None = None) -> Pa
 
     key = jax.random.PRNGKey(cfg.seed)
     fit_key, draw_key = jax.random.split(key)
-    init = jnp.asarray(cfg.initial_position, dtype=jnp.float64)
+    init = jnp.asarray(cfg.initial_position, dtype=jnp.float32)
 
     state, info = pathfinder.approximate(
         fit_key,
